+++ packages/expo-video/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- [Android] Improve audio focus management ([#28453](https://github.com/expo/expo/pull/28453) by [@behenate](https://github.com/behenate))
+
 ## 1.1.3 — 2024-04-24
 
 _This version does not introduce any user-facing changes._